[AeroSpace] Making Service Mode more visible
This probably isn't the best way of doing it, but it's the way I
found! First, when entering the service mode you change the window
borders to a screaming color, and when moving back to mode main you
restore to your default confituration with the borders only in active
windows and stuff like that.